data-collection/requirements.txt
requests
pydantic
python-dateutil
beautifulsoup4
pdfkit
pdfplumber
openai
pyyaml
tiktoken
